Move Annotation class to cpp file
